requirements.txt
# ---------------------------------------------
# 📊 DATA PROCESSING & ANALYSIS
# ---------------------------------------------
pandas              # Core tabular data manipulation
numpy               # Numerical arrays, vectorised ops
pyarrow             # Fast Arrow/Parquet IO for data/artefacts
python-dotenv       # Load environment variables from .env for local runs

# ---------------------------------------------
# 🧠 MACHINE LEARNING
# ---------------------------------------------
scikit-learn        # Pipelines, preprocessing, model selection & metrics
xgboost             # High-performance gradient boosting for tabular data
category-encoders   # Useful encoders (e.g., target/ordinal/WOE) for tabular features

# ---------------------------------------------
# 📈 VISUALIZATION
# ---------------------------------------------
matplotlib          # Base plotting (metrics, diagnostics)
seaborn             # Quick statistical visuals built on matplotlib

# ---------------------------------------------
# 📦 EXPERIMENT TRACKING & MODEL MANAGEMENT
# ---------------------------------------------
mlflow              # Experiment tracking, params/metrics, artefacts, model registry

# ---------------------------------------------
# ✅ TESTING
# ---------------------------------------------
pytest              # Unit/integration tests for data & pipeline logic
pytest-cov          # Coverage reporting (used locally and in CI)

# ---------------------------------------------
# ⚡ API DEVELOPMENT (FOR MODEL SERVING)
# ---------------------------------------------
fastapi             # Lightweight/high-perf REST API for inference
uvicorn             # ASGI server to run FastAPI apps
pydantic            # Request/response schema validation for the API

# ---------------------------------------------
# 🧰 TASK RUNNER / AUTOMATION
# ---------------------------------------------
invoke              # Local task runner (e.g., invoke test, cov, clean)

# ---------------------------------------------
# 🛠️ MISCELLANEOUS UTILITIES
# ---------------------------------------------
pyyaml              # YAML config read/write (pipeline configs)
joblib              # Model/transformer serialisation & simple parallelism
setuptools          # Packaging/build metadata
ipykernel           # Jupyter kernel for notebooks

# ---------------------------------------------
# 🧹 CODE QUALITY (DEV-ONLY)
# ---------------------------------------------
black               # Opinionated code formatter
ruff                # Fast linter (flake8/isort rules)
